Make 'FindOrd' return an enum
The "snap" value is only relevant when the inequality checks evaluate to true. The new enum type is basically the boolean value with a snap value optionally attached to the "true" case.

I think it makes it a little clearer in the code what this snap value is there for - and saves a couple of variables in 'find'.