Add example CLI and fix 'find_range'
Implementing 'find_range' correctly turns out to be surprisingly tricky when considering "snap out" elements: The key is to expand the search space to always include all elements that could match via snapping. It seems like there may be a little room for optimization by eliminating a few redundant lookups. Suggestive comments have been added for looking into that later. There are lots of cases to consider so it isn't clear that much can be done.

The example simple CLI can search for a range of records provided on stdin. The records consist of any number of key=value mappings. The CLI args '--from' and '--to' set the constraints as similar mapping pairs and support both "snap in" (using "=") and "snap out" (using "~"). That is, "snap in" finds only values strictly within the (closed) range, "snap out" finds the first element outside the range target if no exact element is matched.