find_range: Clamp indexes to stay within range
Without the "upper lower" index limitation, it would be possible to check a negative index when looking for the lower target. Limiting the "lower upper" index might save an iteration when looking for the upper target, but not having it probably wouldn't cause errors. So we can't really test that it's necessary - the tests just verify that it doesn't break anything.