Improve docs and correct strange behavior
The behavior was apparently intentional as it was thoroughly tested: It allowed values outside the range to be found. For example, query [l;l] might find l+1 as the lower result because it had "snap up". It also meant that queries [l;u] where u<l might find a result for the upper target.

The changes that aren't directly related to this branch should be moved to its own commit against 'main'.