Ensure that only elements actually in range are found
There was a strange behavior (which was apparently intentional as it was thoroughly tested) that allowed values outside the range to be found. For example, query [l;l] might find l+1 as the "lower" result because it had "snap up". Malformed queries with reversed limits (i.e. [l;u] where l>u) also returned strange results. They still do, but they seem a little less strange at least (see below).

The changes were plucked from the 'combine' branch and also include various minor improvements like cleanup and better docstrings. Also, the trait 'Snap' was moved to the test module as it was actually only used there and in the example CLI.

Finally, we include a bunch of tests as a record of the current behavior of queries with reversed targets. The results of such queries is currently undefined (which is usually bad practice to test), but it probably shouldn't be so the tests ensure that the behavior doesn't change accidentally.